+++ content/_meta.tsx
@@ -1,5 +1,5 @@
 import type { Meta } from 'nextra'
-import { ProductSwitcher } from '../app/components/ProductSwitcher'
+import { ProductSwitcher } from '../app/components/product-switcher'
 import metaLinks from '../src/nav-links'
 
 const meta = {
