+++ content/stream/_meta.tsx
@@ -1,10 +1,11 @@
 import type { Meta } from 'nextra'
-import { ProductSwitcher } from '../../app/components/ProductSwitcher'
+import { ProductSwitcher } from '../../app/components/product-switcher'
 import metaLinks from '../../src/nav-links'
 
 const meta = {
   ...metaLinks,
   _productSwitcher: { type: 'separator', title: <ProductSwitcher /> },
+  _intro: { type: 'separator', title: 'Svix Stream' },
   introduction: 'Introduction',
   'event-types': 'Event Types',
   sinks: 'Sinks',
