+++ package.json
@@ -7,6 +7,9 @@
     "pagefind-vercel": "pagefind --site .next/server/app --output-path public/_pagefind"
   },
   "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^7.2.0",
+    "@fortawesome/free-regular-svg-icons": "^7.2.0",
+    "@fortawesome/react-fontawesome": "^3.3.1",
     "@headlessui/react": "^2.2.10",
     "@heroicons/react": "^2.2.0",
     "next": "^16.2.3",
